second publish float instead of int

+++ exc1/receiver/src/subscriber.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 
 import rospy
-from std_msgs.msg import Int16
+from std_msgs.msg import Int16, Float32
 
 def callback(data):
     q = 0.15
     heard = data.data
     new_number  = heard / q
-    second_pub = rospy.Publisher("/kthfs/result", Int16, queue_size=1)
+    second_pub = rospy.Publisher("/kthfs/result", Float32, queue_size=1)
     second_pub.publish(new_number)
     #rospy.loginfo(new_number)
